Move extensions to their own package
Move logics outside of cmd directories, and group them in a single
package.
