+++ extension/cleaning_extension.go
@@ -1,4 +1,4 @@
-package main
+package extension
 
 import (
 	"fmt"
