+++ extension/kube_auth.go
@@ -1,4 +1,4 @@
-package main
+package extension
 
 import (
 	"context"
